Added some dependency injection
Move Client to become a HttpClient, HttpClient should be a singleton (or close to it.

Added MemoryCache for "isLoggedIn" and "LastPostDateTime"

Added [controller] into the base route of the controllers
This means you don't need to specify those in all routes

Removed IIS Express launch settings

Thinking of a better way to do config but will look at that.